Dev Ideas Rebalance (#95)
Mean time to happen at 0% development cost modifiers to base development is 12 months. Or an average of 50 monarch points/year.

Removed MTTH modifiers involving prosperity and stability because they affect development costs on their own.

Enable is_mtth_scaled_to_size = yes, this will cause MTTH to be scaled up as the number of provinces their owner has increases. This means that the event is just as likely to happen for a large country as for a small one.